# specs/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Stellar Scout API",
		"version": "1.6.1",
		"summary": "Agent-native data layer for the Stellar ecosystem — projects, builders, hackathons, SCF, audits, research.",
...
		"contact": {
			"name": "Stellar Light",
			"url": "https://stellarlight.xyz/scout",
			"email": "support@stellarlight.xyz"
		},
		"license": {
			"name": "MIT",
			"identifier": "MIT"
		}
	},
	"servers": [
		{
			"url": "https://stellarlight.xyz",
			"description": "Production"
		}
	],
...
	"paths": {
...
		"/api/research": {
			"get": {
				"operationId": "searchResearch",
				"tags": [
					"Research"
				],
				"summary": "Vector search over the Stellar research corpus",
				"description": "Semantic ($vectorSearch over Voyage embeddings) search across the Stellar **knowledge corpus** — SDF blog, SCF Handbook, SEPs/standards, dev docs, papers, SCF proposals, security audits, incident reports, and Lumenloop/EC research. Returns the top matching text chunks with source attribution, section, URL, and a confidence score (relevance + freshness + authority); audit chunks add auditor/protocol/severity. Filterable by `source`; falls back to BM25-lite keyword search if vectors are unavailable. **Use when:** 'how does X work', 'is X possible / has X been discussed on Stellar', 'what does the SEP/spec/audit say about X', or you need primary-source citations for a thesis or design question. **Not for:** what products exist or their funding/status → use search_projects; GitHub source code ranked by quality → use search_repos. Use for ecosystem KNOWLEDGE & explainer questions — 'what is / how does / who / why / is it true that…'. Topics: security & risk (bug bounty, vulnerability disclosure, security audit, exploit, hack, incident, post-mortem, oracle manipulation, YieldBlox, Reflector); compliance & regulation (Travel Rule, FATF, KYC, AML, BSA, sanctions, humanitarian aid); funding & governance (SCF v7.0, award tiers, grant programs, Neural Quorum Governance, community voting, quorum); the Stellar Development Foundation (SDF) — mission, legal structure, org; protocol history (Stellar Consensus Protocol, SCP, whitepaper, authors); ecosystem programs (ambassador program, regional chapters, India, bootcamps); and fact-checking or verifying a claim about Stellar. More topics this corpus answers: SCF program mechanics — how to apply (step-by-step application process), Build Award, Instawards, Liquidity Award (amounts), Public Goods Award, eligibility, review timeline (prescreen, panel review, community vote, KYC), submission deadlines; Soroban security incidents — reentrancy, soroban-sdk security advisories/CVEs, denial-of-service (instance-storage growth, authorization recursion); ecosystem history — Soroban mainnet launch (Protocol 20), XLM initial supply creation and distribution, humanitarian aid disbursements (UNHCR Ukraine), SDF Enterprise Fund investments, SDF mandate/spending priorities; making an issued asset tradable/visible on exchanges, wallets, explorers, and aggregators; contract source verification (release.yml, Stellar Lab vs stellar.expert).",
				"parameters": [
					{
						"name": "q",
						"in": "query",
						"required": true,
						"description": "Natural-language search query",
						"schema": {
							"type": "string",
							"minLength": 2
						}
					},
					{
						"name": "source",
						"in": "query",
						"description": "Optional source filter. Use 'audit' for security questions, 'ec-developer-report' for ecosystem stats, 'paper' for foundational protocol questions.",
						"schema": {
							"type": "string",
							"enum": [
								"sdf-blog",
								"scf-handbook",
								"sep",
								"dev-docs",
								"paper",
								"scf-proposal",
								"lumenloop",
								"lumenloop-research",
								"audit",
								"incident",
								"ec-developer-report"
							]
						}
					},
					{
						"name": "limit",
						"in": "query",
						"description": "Max results (default 8, max 25)",
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 25,
							"default": 8
						}
					}
				],
				"responses": {
					"200": {
						"description": "Research results",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								}
							}
						}
					}
				}
			}
		},
...
	},
...
}